# Conflicts:
#	backend/src/main/java/com/ideahub/dao/IdeaDAO.java
#	backend/src/main/java/com/ideahub/resources/IdeaResource.java

+++ backend/src/test/java/com/ideahub/dao/IdeaCollaboratorDAOTest.java
@@ -74,7 +74,7 @@ public class IdeaCollaboratorDAOTest {
 
         IdeaCollaboratorId collabId = IdeaCollaboratorId.builder()
                 .userId((int) userB.getId())
-                .ideaId((int) idea.getId())
+                .ideaId(idea.getId())
                 .build();
         final IdeaInvitation invitation = IdeaInvitation.builder()
                 .acceptedState(false)
